Create README.md
This Code Louisville project is a tribute page to Beyoncé. The required CSS features that are included are a hamburger menu that opens and closes by clicking. There is also an animation that occurs to close out the hamburger menu.  The required JavaScript features are a "Contact Us" form that displays a message once the form is completely filled out and will display an error if any fields are missing or entered incorrectly.